style(yk-color-picker): update light theme colors
Updated the background colors for the light theme in the yk-color-picker library. This includes a change from white (#fff) to slight grey (#f7f7f7) for the color picker wrapper and from white (#fff) to off-white (#fcfcfc) for the color input field. The changes aim to provide a better user experience by reducing brightness in the light theme.