feat(yk-color-picker): improve button handling and keyup event
This commit improves user interaction in the YKColorPicker module. It renames the `inputsSwitch` element to `btnSwitch` for increased clarity and ease of reference. Additionally, it adds a new condition to the `#onKeyUpClose` event. Now the color picker will only close when the "Enter" key is pressed if neither the "Copy Color" nor the "Switch Button" are the targets. This prevents accidental closing of the color picker when interacting with these elements.