feat(package & vite config): update library version and expand build formats
This commit includes changes in two main areas. In the package.json file, the version of the yk-color-picker library was updated from 1.0.9 to 1.0.10. Additionally, a "main" entry point was defined pointing to the UMD (Universal Module Definition) build format of the library.

The vite.config.js file was also modified to include the UMD format in the library's build process, further expanding its compatibility with different module systems. This change allows the library to be consumed not only as an ES module but also as a UMD module, making it more flexible for various use cases.